requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0
gradio>=4.0.0

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# PDF Processing
PyPDF2>=3.0.0

# Data Handling
pandas>=2.0.0

# Utilities
typing-extensions>=4.5.0
